+++ changelog.md
@@ -5,6 +5,7 @@ Internal
 --------
 * Improve test coverage for `completion_refresher.py`.
 * Add test coverage for `client_query.py`.
+* Improve test coverage for `output.py`.
 
 
 1.74.0 (2026/06/06)